+++ packages/core/src/fonts.ts
@@ -318,3 +318,18 @@ export function weightToStyle(weight: number, italic = false): string {
   if (italic) label += ' Italic'
   return label
 }
+
+export function weightToFigmaStyle(weight: number, italic = false): string {
+  let label = 'Regular'
+  if (weight <= 100) label = 'Thin'
+  else if (weight <= 200) label = 'Extra Light'
+  else if (weight <= 300) label = 'Light'
+  else if (weight <= 400) label = 'Regular'
+  else if (weight <= 500) label = 'Medium'
+  else if (weight <= 600) label = 'Semi Bold'
+  else if (weight <= 700) label = 'Bold'
+  else if (weight <= 800) label = 'Extra Bold'
+  else if (weight >= 900) label = 'Black'
+  if (italic) label += ' Italic'
+  return label
+}